Better documentation, docs site, clean up
