+++ .eslintignore
@@ -1,4 +1,3 @@
 .nyc_output
-coverage
-site/themes/*
-convert
+node_modules
+
