+++ .eslintrc
@@ -30,8 +30,7 @@
   "parserOptions": {
     "ecmaVersion": 10
   },
-  # Custom Disabled Rules
-  rules: {
+  "rules": {
     "import/no-dynamic-require": 0,
     "import/no-extraneous-dependencies": [
       "error", {
@@ -48,14 +47,14 @@
     "security/detect-object-injection": 0,
     "unicorn/filename-case": 0,
     "unicorn/prevent-abbreviations": 0,
-    camelcase: 0,
-    global-require: 0,
-    max-len: 0,
-    no-empty: ["error", { "allowEmptyCatch": true }],
-    no-param-reassign: 0,
-    no-plusplus: 0, # Meaningless as other rules prevent the errors that could happen.
-    no-restricted-syntax: 0,
-    no-underscore-dangle: 0,
-    no-unused-vars: ["error", { "varsIgnorePattern": "^_", "argsIgnorePattern": "^_", "caughtErrors": "none" }]
+    "camelcase": 0,
+    "global-require": 0,
+    "max-len": 0,
+    "no-empty": ["error", { "allowEmptyCatch": true }],
+    "no-param-reassign": 0,
+    "no-plusplus": 0,
+    "no-restricted-syntax": 0,
+    "no-underscore-dangle": 0,
+    "no-unused-vars": ["error", { "varsIgnorePattern": "^_", "argsIgnorePattern": "^_", "caughtErrors": "none" }]
   }
 }
