--- .tern-project
@@ -1,11 +1,0 @@
-{
-  "ecmaVersion": 6,
-  "libs": [],
-  "loadEagerly": [
-    "app/**/*.js"
-  ],
-  "plugins": {
-    "node": {},
-    "es_modules": {}
-  }
-}
