+++ README.hbs
@@ -1,0 +1,20 @@
+[![view on npm](http://img.shields.io/npm/v/uttori-search-provider-lunr.svg)](https://www.npmjs.org/package/uttori-search-provider-lunr)
+[![npm module downloads](http://img.shields.io/npm/dt/uttori-search-provider-lunr.svg)](https://www.npmjs.org/package/uttori-search-provider-lunr)
+[![Build Status](https://travis-ci.org/uttori/uttori-search-provider-lunr.svg?branch=master)](https://travis-ci.org/uttori/uttori-search-provider-lunr)
+[![Dependency Status](https://david-dm.org/uttori/uttori-search-provider-lunr.svg)](https://david-dm.org/uttori/uttori-search-provider-lunr)
+[![Coverage Status](https://coveralls.io/repos/uttori/uttori-search-provider-lunr/badge.svg?branch=master)](https://coveralls.io/r/uttori/uttori-search-provider-lunr?branch=master)
+
+#  Uttori Search Provider - Lunr
+Uttori Search Provider powered by [Lunr.js](https://lunrjs.com/).
+
+## Install
+
+```
+$ npm install --save uttori-search-provider-lunr
+```
+
+# API Reference
+
+{{>main}}
+
+* * *
