+++ ava.config.js
@@ -1,9 +1,10 @@
+/* eslint-disable node/no-unsupported-features/es-syntax */
 export default {
   files: [
     'test/**/*.test.js',
   ],
   sources: [
-    'index.js',
+    'src/index.js',
   ],
   concurrency: 5,
   failFast: false,
