+++ package.json
@@ -22,6 +22,7 @@
   },
   "devDependencies": {
     "ava": "^1.4.1",
+    "docsify-cli": "^4.3.0",
     "eslint": "^5.16.0",
     "eslint-config-airbnb": "^17.1.0",
     "eslint-plugin-ava": "^6.0.0",
@@ -35,12 +36,12 @@
     "eslint-plugin-security": "^1.4.0",
     "eslint-plugin-unicorn": "^8.0.2",
     "eslint-plugin-xss": "^0.1.9",
+    "jsdoc-to-markdown": "^4.0.1",
     "nyc": "^14.0.0",
-    "sinon": "^7.3.2",
-    "supertest": "^4.0.2"
+    "sinon": "^7.3.2"
   },
   "files": [
-    "index.js"
+    "src/index.js"
   ],
   "main": "uttori-search-provider-lunr",
   "nyc": {
@@ -48,11 +49,13 @@
     "all": true,
     "check-coverage": true,
     "include": [
-      "index.js"
+      "src/index.js"
     ],
     "exclude": []
   },
   "scripts": {
+    "docs": "jsdoc2md --private --example-lang js --template README.hbs src/index.js > README.md; cp README.md docs/README.md",
+    "docs-serve": "docsify serve ./docs",
     "lint": "eslint .",
     "report": "nyc report --reporter=html",
     "test": "NODE_ENV=test nyc ava",
